replace metrics logging with FedValidator and restructure validator module
