+++ froseai/validator/__init__.py
@@ -1,0 +1,5 @@
+from .valid_mixin import FedValidator
+
+__all__ = [
+    "FedValidator"
+]
